move Item inside of Assets
